+++ doers/src/zone/illumos.rs
@@ -1,29 +1,34 @@
-use crate::zone::constants::OMNIOS_RELEASES_URL;
+use crate::zone::helpers;
+use crate::zone::types::ZoneImage;
+use crate::zone::{config::ImageChecksum, constants::OMNIOS_RELEASES_URL};
 use anyhow::{Context, bail};
 use camino::Utf8PathBuf;
-use common::constants::IMG_CACHE_DIR;
-use common::types::ApplyOpts;
-use util::http;
 
 // Images follow the form:
 // https://downloads.omnios.org/media/stable/omnios-r151056.ngz.zfs.xz
 // For now we will simply download the stable NGZ image
 
-pub fn image_path(image: Option<&str>) -> anyhow::Result<Utf8PathBuf> {
+pub fn image_path(image: ZoneImage) -> anyhow::Result<Utf8PathBuf> {
     // if we're given an image and it looks like a URL, fetch it. If it looks like a file, stat
     // it; if we don't have one, fetch the current stable
     // if let Some() = image {
-    match image {
-        Some(path) => {
-            if path.starts_with("/") {
-                Ok(Utf8PathBuf::from(path))
-            } else if path.starts_with("http") {
-                get_image(path)
+    match image.user_string {
+        Some(user_string) => {
+            if user_string.starts_with("/") {
+                Ok(Utf8PathBuf::from(user_string))
+            } else if user_string.starts_with("http") {
+                helpers::get_image(user_string, image.checksum)
             } else {
                 bail!("illumos image must be a fully qualified path or URL")
             }
         }
-        None => get_image(&default_image()?),
+        None => helpers::get_image(
+            &default_image()?,
+            Some(&ImageChecksum {
+                sumtype: "sha256".into(),
+                value: ".sha256".into(),
+            }),
+        ),
     }
 }
 
@@ -51,24 +56,5 @@ fn default_image() -> anyhow::Result<String> {
         Ok(image_url)
     } else {
         bail!("could not find ngz.zfs.xz image");
-    }
-}
-
-fn get_image(img_url: &str) -> anyhow::Result<Utf8PathBuf> {
-    let chunks = img_url.split("/");
-
-    if let Some(img_basename) = chunks.last() {
-        let img_path = Utf8PathBuf::from(IMG_CACHE_DIR).join(img_basename);
-
-        if img_path.exists() {
-            tracing::debug!("found image at {img_path}");
-        } else {
-            tracing::debug!("no image at {img_path}: downloading");
-            http::remote_file_to_disk(img_url, &img_path, None, &ApplyOpts::default())?;
-        }
-
-        Ok(img_path)
-    } else {
-        bail!("could not get image basename");
     }
 }